+++ docs/tasks.js
@@ -53,6 +53,7 @@
  * @description ESLint configuration object, can contain any options of the eslint cli engine.
  * @see http://eslint.org/docs/developer-guide/nodejs-api#cliengine
  * @property {string|string[]} files Paths of the files to check
+ * @property {string[]} extensions File extensions to validate
  * @property {string} configFile The path to the eslint configuration file
  */
 
@@ -60,6 +61,7 @@
  * @typedef {Object|function} Parameters
  * @memberof module:tasks/jscs
  * @property {string|string[]} files Paths of the files to check
+ * @property {string[]} extensions File extensions to validate
  * @property {string} configFile The path to the jscs configuration file
  */