+++ src/tasks/eslint/__tests__/file-extensions/gulpfile.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const gulp = require('gulp');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'eslint', {
+  configFile: '.eslintrc',
+  files: '**/*',
+  extensions: ['.jsx']
+});