+++ src/tasks/jscs/__tests__/file-extensions/gulpfile.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const gulp = require('gulp');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'jscs', {
+  configFile: '.jscsrc',
+  files: '**/*',
+  extensions: ['.jsx']
+});