+++ src/tasks/jscs/__tests__/valid-project/index.js
@@ -1,2 +1,1 @@
-export function valid() {
-}
+export function valid() {}