+++ src/tasks/jscs/defaults.js
@@ -4,5 +4,6 @@ const path = require('path');
 
 module.exports = {
   configFile: path.join(__dirname, '../../../.jscsrc'),
+  extensions: ['.js'],
   files: require('../../utils/global-source-files')
 };